use caller-specified domain instead of server-returned target name in Type 3 Message
According to Eric Glass's NTLM Authentication protocol
documentation at http://davenport.sourceforge.net/ntlm.html,
the Target Name field in the Type 3 Message should contain
the NT domain name when the DOMAIN\user name variation is
used or should be empty when Kerberos-style "user@realm"
format is used. This change uses the caller specified domain
argument instead of the target name returned in the
Type 2 Message. The caller can omit the domain name when
using the "user@DOMAIN" or "user@domain.com" username
and specify it when using the "DOMAIN\user" or
"domain.com\user" format. Without this change the Type 3
Message Target Name field would not necessarily be empty
when the former username format was used and the Type 2
Message returned a target name.