Avoid usage of legacy algorithms on libssl-3.0+
Single DES and MD4 are considered legacy algorithms in OpenSSL/libssl-3.0.
They can be enabled by adjusting the openssl configuration file or by using the new provider concept of OpenSSL-3.
Editing the configuration must be done by each use of rubyntlm, which is very inconvenient.
The provider API is not yet supported by the ruby binding to OpenSSL.

So it's better to avoid the legacy algorithms at all.
The single DES algorithm can easily implemented by a two key 3-DES run.
The md4 implementation is taken from here:
  https://gist.github.com/tprynn/5419da1a2ad8935c1fff
And the rc4 implementation is taken from here and modified:
  https://github.com/caiges/Ruby-RC4/blob/082fce56ab707dc77442709357cf176f7b3b6f22/lib/rc4.rb

Fixes #49